MODIFY : slide component 생성
