+++ src/components/Main/SlideList.js
